Add mp4 and YouTube videos support

+++ src/pages/api/photo/deletePhoto.js
@@ -4,6 +4,15 @@ import prisma from "../../../../utils/prisma"
 export default async function deletePhoto(req, res) {
     try {
         const photo = await prisma.photo.findFirst({
+            select: {
+                id: true,
+                link: true,
+                Category: {
+                    select: {
+                        type: true
+                    }
+                }
+            },
             where: {
                 link: req.body.link,
             },
@@ -15,7 +24,9 @@ export default async function deletePhoto(req, res) {
             }
         })
 
-        fs.unlinkSync(`./public/${photo.link}`)
+        if (photo.Category.type !== 'youtube') {
+            fs.unlinkSync(`./public/${photo.link}`)
+        }
 
         res.status(200).json({content: {}})
     } catch (err) {
